+++ oggm/cfg.py
@@ -46,6 +46,7 @@ CONFIG_MODIFIED = False
 # Share state accross processes
 DL_VERIFIED = dict()
 DEM_SOURCE_TABLE = dict()
+DATA = dict()
 
 # Machine epsilon
 FLOAT_EPS = np.finfo(float).eps
@@ -165,7 +166,6 @@ PARAMS = ParamsLoggingDict()
 PATHS = PathOrderedDict()
 BASENAMES = DocumentedDict()
 LRUHANDLERS = ResettingOrderedDict()
-DATA = ResettingOrderedDict()
 
 # Constants
 SEC_IN_YEAR = 365*24*3600
@@ -538,6 +538,7 @@ def initialize_minimal(file=None, logging_level='INFO', params=None,
     PARAMS['climate_qc_months'] = cp.as_int('climate_qc_months')
     PARAMS['temp_use_local_gradient'] = cp.as_bool('temp_use_local_gradient')
     PARAMS['tstar_search_glacierwide'] = cp.as_bool('tstar_search_glacierwide')
+    PARAMS['geodetic_mb_period'] = cp['geodetic_mb_period']
 
     k = 'temp_local_gradient_bounds'
     PARAMS[k] = [float(vk) for vk in cp.as_list(k)]
@@ -583,7 +584,7 @@ def initialize_minimal(file=None, logging_level='INFO', params=None,
            'error_when_glacier_reaches_boundaries', 'glacier_length_method',
            'use_inversion_params_for_run', 'ref_mb_valid_window',
            'tidewater_type', 'store_model_geometry',
-           'store_diagnostic_variables']
+           'store_diagnostic_variables', 'geodetic_mb_period']
     for k in ltr:
         cp.pop(k, None)
 
@@ -816,7 +817,6 @@ def pack_config():
         'LRUHANDLERS': LRUHANDLERS,
         'DATA': DATA,
         'BASENAMES': dict(BASENAMES),
-
         'DL_VERIFIED': DL_VERIFIED,
         'DEM_SOURCE_TABLE': DEM_SOURCE_TABLE
     }
@@ -833,6 +833,8 @@ def unpack_config(cfg_dict):
     PATHS = cfg_dict['PATHS']
     LRUHANDLERS = cfg_dict['LRUHANDLERS']
     DATA = cfg_dict['DATA']
+    DL_VERIFIED = cfg_dict['DL_VERIFIED']
+    DEM_SOURCE_TABLE = cfg_dict['DEM_SOURCE_TABLE']
 
     # BASENAMES is a DocumentedDict, which cannot be pickled because
     # set intentionally mismatches with get
@@ -840,14 +842,11 @@ def unpack_config(cfg_dict):
     for k in cfg_dict['BASENAMES']:
         BASENAMES[k] = (cfg_dict['BASENAMES'][k], 'Imported Pickle')
 
-    DL_VERIFIED = cfg_dict['DL_VERIFIED']
-    DEM_SOURCE_TABLE = cfg_dict['DEM_SOURCE_TABLE']
-
 
 def set_manager(manager):
     """Sets a multiprocessing manager to use for shared dicts"""
 
-    global DL_VERIFIED, DEM_SOURCE_TABLE
+    global DL_VERIFIED, DEM_SOURCE_TABLE, DATA
 
     if manager:
         new_dict = manager.dict()
@@ -857,9 +856,14 @@ def set_manager(manager):
         new_dict = manager.dict()
         new_dict.update(DEM_SOURCE_TABLE)
         DEM_SOURCE_TABLE = new_dict
+
+        new_dict = manager.dict()
+        new_dict.update(DATA)
+        DATA = new_dict
     else:
         DL_VERIFIED = dict(DL_VERIFIED)
         DEM_SOURCE_TABLE = dict(DEM_SOURCE_TABLE)
+        DATA = dict(DATA)
 
 
 def add_to_basenames(basename, filename, docstr=''):