+++ oggm/shop/ecmwf.py
@@ -116,11 +116,6 @@ def process_ecmwf_data(gdir, dataset=None, ensemble_member=0,
         previous experiments)
     """
 
-    if cfg.PATHS.get('climate_file', None):
-        warnings.warn("You seem to have set a custom climate file for this "
-                      "run, but are using the ECMWF climate file "
-                      "instead.")
-
     if dataset is None:
         dataset = cfg.PARAMS['baseline_climate']
 
@@ -146,7 +141,7 @@ def process_ecmwf_data(gdir, dataset=None, ensemble_member=0,
             ds = ds.sel(number=ensemble_member)
         try:
             ds = ds.sel(longitude=lon, latitude=lat, method='nearest')
-        except ValueError:
+        except (ValueError, KeyError):
             # Flattened ERA5
             c = (ds.longitude - lon)**2 + (ds.latitude - lat)**2
             ds = ds.isel(points=np.argmin(c.data))
@@ -163,7 +158,7 @@ def process_ecmwf_data(gdir, dataset=None, ensemble_member=0,
             ds = ds.sel(number=ensemble_member)
         try:
             ds = ds.sel(longitude=lon, latitude=lat, method='nearest')
-        except ValueError:
+        except (ValueError, KeyError):
             # Flattened ERA5
             c = (ds.longitude - lon)**2 + (ds.latitude - lat)**2
             ds = ds.isel(points=np.argmin(c.data))
@@ -173,7 +168,7 @@ def process_ecmwf_data(gdir, dataset=None, ensemble_member=0,
         ds = ds.isel(time=0)
         try:
             ds = ds.sel(longitude=lon, latitude=lat, method='nearest')
-        except ValueError:
+        except (ValueError, KeyError):
             # Flattened ERA5
             c = (ds.longitude - lon)**2 + (ds.latitude - lat)**2
             ds = ds.isel(points=np.argmin(c.data))